Move is_between to utils/dataframe
